add random point test, debug_eval pass
